+++ hyp3_gamma/__init__.py
@@ -9,6 +9,6 @@ try:
 except PackageNotFoundError:
     print('package is not installed!\n'
           'Install in editable/develop mode via (from the top of this repo):\n'
-          '   python -m pip install -e .\n'
+          '   python -m pip install -e .[develop]\n'
           'Or, to just get the version number use:\n'
           '   python setup.py --version')
