+++ hyp3_gamma/rtc/smoothem.py
